+++ src/main/java/autopartner/controller/rest/OrderTaskController.java
@@ -35,7 +35,7 @@ public class OrderTaskController {
 
     @Secured({"ROLE_ADMIN", "ROLE_ROOT"})
     @RequestMapping(value = "/api/order/task/{id}", method = RequestMethod.GET)
-    public ResponseEntity<OrderTask> get(@PathVariable Integer id) {
+    public ResponseEntity<OrderTask> get(@PathVariable Long id) {
         return ResponseEntity.ok(orderTaskService.getOrderTaskById(id));
     }
