+++ src/main/java/autopartner/repository/CarRepository.java
@@ -3,6 +3,6 @@ package autopartner.repository;
 import autopartner.domain.entity.Car;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarRepository extends CrudRepository<Car, Integer> {
+public interface CarRepository extends CrudRepository<Car, Long> {
     Iterable<Car> findByActiveTrue();
 }
