+++ src/main/java/autopartner/repository/MaterialTypeRepository.java
@@ -3,6 +3,6 @@ package autopartner.repository;
 import autopartner.domain.entity.MaterialType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MaterialTypeRepository extends CrudRepository<MaterialType, Integer> {
+public interface MaterialTypeRepository extends CrudRepository<MaterialType, Long> {
     Iterable<MaterialType> findByActiveTrue();
 }
